+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Widgets - Improve UX of custom subtitle documentation (@codewithdipesh)
 - Main screen - Show “Negligible” inside Pollen block if there is no pollen today instead of an empty block
 
+**Weather sources**
+- [GeoSphere AT] Fix missing info in warnings
+
 **Translations**
 - Translations updated